Requested changes from code review
